# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AnmeldungInputBezahlweg =
  (typeof AnmeldungInputBezahlweg)[keyof typeof AnmeldungInputBezahlweg];

export const AnmeldungInputBezahlweg = {
  ueberweisung: "ueberweisung",
  paypal: "paypal",
} as const;

export interface AnmeldungInput {
  email: string;
  /** @nullable */
  telefon?: string | null;
  /**
   * @minimum 1
   * @maximum 6
   */
  personen_anzahl: number;
  personen: string[];
  bezahlweg?: AnmeldungInputBezahlweg;
  /** @nullable */
  song?: string | null;
  /**
   * @maxLength 200
   * @nullable
   */
  statement?: string | null;
  /** Muss true sein */
  verbindlich: boolean;
}

export interface AnmeldungResult {
  id: number;
  betrag_gesamt: number;
  /** Fortlaufende Einlass-Nummern, eine pro Person */
  ticket_nummern: number[];
}

export interface AnmeldungStats {
  angemeldete_personen: number;
}

export interface InteressentenCount {
  count: number;
}

export interface InteresseStats {
  /** Anzahl Phase-1-Interessenten */
  boomer: number;
  /** Summe der angemeldeten Personen aus Phase-1 */
  personen: number;
}

export interface InteresseEntry {
  id: string;
  name: string;
  /** @nullable */
  personen?: string | null;
  /** @nullable */
  statement?: string | null;
  /** @nullable */
  song?: string | null;
  /** @nullable */
  createdAt?: string | null;
  /**
   * Vom Admin freigegebener Anzeigename (oder null wenn noch ausstehend)
   * @nullable
   */
  display_name?: string | null;
}

export interface InteresseResponse {
  stats: InteresseStats;
  entries: InteresseEntry[];
}

export interface DisplayNameEntry {
  id: number;
  source_type: string;
  source_id: string;
  raw_name: string;
  song: string;
  suggested_name: string;
  /** @nullable */
  approved_name?: string | null;
  status: string;
  /** @nullable */
  updated_at?: string | null;
}

export type DisplayNamePatchStatus =
  (typeof DisplayNamePatchStatus)[keyof typeof DisplayNamePatchStatus];

export const DisplayNamePatchStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface DisplayNamePatch {
  status?: DisplayNamePatchStatus;
  /** @nullable */
  approved_name?: string | null;
}

export interface DisplayNamesSyncResult {
  synced: number;
}

export interface StoRnierungResult {
  ok: boolean;
  storniert_am: string;
}

export interface OkResult {
  ok: boolean;
}

export type ApiErrorDetails = { [key: string]: unknown };

export interface ApiError {
  error: string;
  details?: ApiErrorDetails;
}