# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: anmeldung
    description: Event registration
  - name: interesse
    description: Interessensbekundungen
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /interesse:
    get:
      operationId: getInteresse
      tags: [interesse]
      summary: Interessensbekundungen und Songwünsche abrufen
      description: Liefert Phase-1-Stats (nur interessenten-Tabelle) sowie eine gemischte Eintrags-Liste für die Anzeige
      responses:
        "200":
          description: Interesse-Daten
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteresseResponse"
        "500":
          description: Serverfehler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /interessenten/count:
    get:
      operationId: getInteressentenCount
      tags: [interesse]
      summary: Anzahl der Interessensbekundungen
      description: Gibt die Gesamtzahl der Einträge in der interessenten-Tabelle zurück
      responses:
        "200":
          description: Anzahl
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteressentenCount"
        "500":
          description: Serverfehler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /anmeldung:
    post:
      operationId: submitAnmeldung
      tags: [anmeldung]
      summary: Verbindliche Anmeldung einreichen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnmeldungInput"
      responses:
        "201":
          description: Anmeldung erfolgreich gespeichert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnmeldungResult"
        "400":
          description: Validierungsfehler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Serverfehler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/display-names/sync:
    post:
      operationId: syncDisplayNames
      tags: [interesse]
      summary: Neue Songwunsch-Namen einlesen
      description: Legt für alle Interessenten/Anmeldungen mit Song fehlende display_names-Zeilen an
      parameters:
        - in: header
          name: x-admin-secret
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Anzahl neu angelegter Zeilen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisplayNamesSyncResult"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/display-names:
    get:
      operationId: listDisplayNames
      tags: [interesse]
      summary: Alle Anzeigenamen-Einträge abrufen
      parameters:
        - in: header
          name: x-admin-secret
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Liste der Einträge
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DisplayNameEntry"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/display-names/{id}:
    patch:
      operationId: updateDisplayName
      tags: [interesse]
      summary: Anzeigenamen-Eintrag aktualisieren
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: header
          name: x-admin-secret
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisplayNamePatch"
      responses:
        "200":
          description: Erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /anmeldung/stats:
    get:
      operationId: getAnmeldungStats
      tags: [anmeldung]
      summary: Anmeldungsstatistik abrufen
      description: Liefert die Gesamtzahl angemeldeter Personen (stornierte ausgeschlossen)
      responses:
        "200":
          description: Statistik
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnmeldungStats"
        "500":
          description: Serverfehler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/anmeldungen/{id}/stornieren:
    post:
      operationId: stornierenAnmeldung
      tags: [anmeldung]
      summary: Anmeldung intern stornieren
      description: Setzt storniert_am auf now(). Stille Aktion — kein Mail an Teilnehmer.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: header
          name: x-admin-secret
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Erfolgreich storniert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoRnierungResult"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Nicht gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/anmeldungen/{id}/reaktivieren:
    post:
      operationId: reaktivierenAnmeldung
      tags: [anmeldung]
      summary: Stornierung einer Anmeldung rückgängig machen
      description: Setzt storniert_am zurück auf null. Stille Aktion — kein Mail an Teilnehmer.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: header
          name: x-admin-secret
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Erfolgreich reaktiviert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Nicht gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AnmeldungInput:
      type: object
      required:
        - email
        - personen_anzahl
        - personen
        - verbindlich
      properties:
        email:
          type: string
          format: email
        telefon:
          type: ["string", "null"]
        personen_anzahl:
          type: integer
          minimum: 1
          maximum: 6
        personen:
          type: array
          items:
            type: string
            minLength: 2
        bezahlweg:
          type: string
          enum: [ueberweisung, paypal]
        song:
          type: ["string", "null"]
        statement:
          type: ["string", "null"]
          maxLength: 200
        verbindlich:
          type: boolean
          description: Muss true sein
    AnmeldungResult:
      type: object
      required:
        - id
        - betrag_gesamt
        - ticket_nummern
      properties:
        id:
          type: integer
        betrag_gesamt:
          type: integer
        ticket_nummern:
          type: array
          items:
            type: integer
          description: Fortlaufende Einlass-Nummern, eine pro Person
    AnmeldungStats:
      type: object
      required:
        - angemeldete_personen
      properties:
        angemeldete_personen:
          type: integer
    InteressentenCount:
      type: object
      required:
        - count
      properties:
        count:
          type: integer
    InteresseStats:
      type: object
      required:
        - boomer
        - personen
      properties:
        boomer:
          type: integer
          description: Anzahl Phase-1-Interessenten
        personen:
          type: integer
          description: Summe der angemeldeten Personen aus Phase-1
    InteresseEntry:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        personen:
          type: ["string", "null"]
        statement:
          type: ["string", "null"]
        song:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
        display_name:
          type: ["string", "null"]
          description: Vom Admin freigegebener Anzeigename (oder null wenn noch ausstehend)
    InteresseResponse:
      type: object
      required:
        - stats
        - entries
      properties:
        stats:
          $ref: "#/components/schemas/InteresseStats"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/InteresseEntry"
    DisplayNameEntry:
      type: object
      required:
        - id
        - source_type
        - source_id
        - raw_name
        - song
        - suggested_name
        - status
      properties:
        id:
          type: integer
        source_type:
          type: string
        source_id:
          type: string
        raw_name:
          type: string
        song:
          type: string
        suggested_name:
          type: string
        approved_name:
          type: ["string", "null"]
        status:
          type: string
        updated_at:
          type: ["string", "null"]
    DisplayNamePatch:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
        approved_name:
          type: ["string", "null"]
    DisplayNamesSyncResult:
      type: object
      required:
        - synced
      properties:
        synced:
          type: integer
    StoRnierungResult:
      type: object
      required:
        - ok
        - storniert_am
      properties:
        ok:
          type: boolean
        storniert_am:
          type: string
          format: date-time
    OkResult:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
    ApiError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details:
          type: object